# contracts/openapi.yaml
components:
  schemas:
...
    ErrorResponse:
      description: Stable error envelope returned by every ULOO Core endpoint.
      properties:
        code:
          description: Stable machine-readable error code
          title: Code
          type: string
        details:
          anyOf:
          - {}
          - type: 'null'
          title: Details
        message:
          description: Safe human-readable error message
          title: Message
          type: string
        request_id:
          title: Request Id
          type: string
        trace_id:
          title: Trace Id
          type: string
      required:
      - code
      - message
      - request_id
      - trace_id
      title: ErrorResponse
      type: object
...
    Page_TeamResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TeamResponse'
          title: Items
          type: array
        limit:
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          minimum: 0.0
          title: Offset
          type: integer
        total:
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - offset
      - limit
      title: Page[TeamResponse]
      type: object
...
    TeamCreate:
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        enabled:
          default: true
          title: Enabled
          type: boolean
        instructions:
          items:
            type: string
          title: Instructions
          type: array
        key:
          description: Unique stable key
          maxLength: 128
          minLength: 1
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
          title: Key
          type: string
        leader_agent_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Required for coordinate, route
          title: Leader Agent Id
        limits:
          $ref: '#/components/schemas/TeamLimits'
        member_agent_ids:
          items:
            format: uuid
            type: string
          title: Member Agent Ids
          type: array
        memory_policy:
          $ref: '#/components/schemas/MemoryPolicy'
        mode:
          description: 'One of: coordinate, route, collaborate'
          title: Mode
          type: string
        name:
          maxLength: 256
          minLength: 1
          title: Name
          type: string
      required:
      - key
      - name
      - mode
      title: TeamCreate
      type: object
    TeamLimits:
      properties:
        max_iterations:
          default: 8
          maximum: 100.0
          minimum: 1.0
          title: Max Iterations
          type: integer
        max_tokens:
          default: 20000
          maximum: 2000000.0
          minimum: 1.0
          title: Max Tokens
          type: integer
        timeout_seconds:
          default: 120
          maximum: 3600.0
          minimum: 1.0
          title: Timeout Seconds
          type: integer
      title: TeamLimits
      type: object
    TeamResponse:
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        enabled:
          title: Enabled
          type: boolean
        id:
          title: Id
          type: string
        instructions:
          items:
            type: string
          title: Instructions
          type: array
        key:
          description: Unique stable key
          maxLength: 128
          minLength: 1
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
          title: Key
          type: string
        leader_agent_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Required for coordinate, route
          title: Leader Agent Id
        limits:
          $ref: '#/components/schemas/TeamLimits'
        member_agent_ids:
          items:
            format: uuid
            type: string
          title: Member Agent Ids
          type: array
        memory_policy:
          $ref: '#/components/schemas/MemoryPolicy'
        mode:
          description: 'One of: coordinate, route, collaborate'
          title: Mode
          type: string
        name:
          maxLength: 256
          minLength: 1
          title: Name
          type: string
        updated_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated At
        version:
          title: Version
          type: integer
      required:
      - key
      - name
      - mode
      - id
      - enabled
      - version
      title: TeamResponse
      type: object
...
info:
  title: ULOO Core
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/teams:
    get:
      description: List and search teams.
      operationId: list_teams_api_v1_teams_get
      parameters:
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - in: query
        name: mode
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Mode
      - in: query
        name: enabled
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_TeamResponse_'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: List Teams
      tags:
      - teams
    post:
      description: Create a new Team definition.
      operationId: create_team_api_v1_teams_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Create Team
      tags:
      - teams
  /api/v1/teams/by-key/{team_key}:
    get:
      description: Get team by stable key (used by plugin).
      operationId: get_team_by_key_api_v1_teams_by_key__team_key__get
      parameters:
      - in: path
        name: team_key
        required: true
        schema:
          title: Team Key
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Get Team By Key
      tags:
      - teams
  /api/v1/teams/{team_id}:
    delete:
      description: Soft delete team.
      operationId: delete_team_api_v1_teams__team_id__delete
      parameters:
      - in: path
        name: team_id
        required: true
        schema:
          format: uuid
          title: Team Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Delete Team
      tags:
      - teams
    get:
      description: Get team details by ID.
      operationId: get_team_api_v1_teams__team_id__get
      parameters:
      - in: path
        name: team_id
        required: true
        schema:
          format: uuid
          title: Team Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Get Team
      tags:
      - teams
    patch:
      description: Update team with optimistic locking.
      operationId: update_team_api_v1_teams__team_id__patch
      parameters:
      - in: path
        name: team_id
        required: true
        schema:
          format: uuid
          title: Team Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Update Team
      tags:
      - teams
...